streamlit_requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-socketio[client]>=5.10.0
pandas>=2.0.0
